Extract edit and save to respective components
