+++ src/blocks/block-testimonial/components/testimonial.js
@@ -7,7 +7,6 @@ const { Component } = wp.element;
 
 // Import block dependencies and components
 import classnames from 'classnames';
-import * as fontSize from './../../../utils/helper';
 
 /**
  * Create a Testimonial wrapper Component
@@ -21,7 +20,7 @@ export default class Testimonial extends Component {
 	render() {
 
 		// Setup the attributes
-		const { attributes: { testimonialAlignment, testimonialImgURL, testimonialBackgroundColor, testimonialTextColor, testimonialFontSize, testimonialCiteAlign }  } = this.props;
+		const { attributes: { testimonialImgURL, testimonialBackgroundColor, testimonialTextColor, testimonialFontSize, testimonialCiteAlign }  } = this.props;
 
 		return (
 			<div
